CredentialsVC: Split platform-specifics into a separate file
